numpy json encoding, move write_log utility to likelihood subpackage.
